Added BANKATM.py file
Added bankatm.py file
An ATM Software In Python is a simple console-based ATM simulator that provides the simple account balance management of a respective account. It contains all the essential features. There is no database connection or neither any external text nor other files used in this mini project to save the user’s data. Everything is set inside the source code whether it's the pin code or the amount.